+++ src/sasctl/core.py
@@ -749,11 +749,49 @@ def delete(path, **kwargs):
     return request('delete', path, **kwargs)
 
 
-def request(verb, path, session=None, raw=False, **kwargs):
+def request(verb, path, session=None, raw=False, format='auto', **kwargs):
+    """Send an HTTP request with a session.
+
+    Parameters
+    ----------
+    verb : str
+        A valid HTTP request verb.
+    path : str
+        Path portion of URL to request.
+    session : Session, optional
+        Defaults to `current_session()`.
+    raw : bool
+        Deprecated. Whether to return the raw `Response` object.
+        Defaults to False.
+    format : {'auto', 'response', 'content', 'json', 'text'}
+        The format of the return response.  Defaults to `auto`.
+        response: the raw `Response` object.
+        content: Response.content
+        json: Response.json()
+        text: Response.text
+        auto: `RestObj` constructed from JSON if possible, otherwise same as
+              `text`.
+    kwargs : any
+        Additional arguments are passed to the session `request` method.
+
+    Returns
+    -------
+
+    """
     session = session or current_session()
 
     if session is None:
         raise TypeError('No `Session` instance found.')
+
+    if raw:
+        warnings.warn("The 'raw' parameter is deprecated and will be removed in"
+                      " a future version.  Use format='response' instead.",
+                      DeprecationWarning)
+        format = 'response'
+
+    format = 'auto' if format is None else str(format).lower()
+    if format not in ('auto', 'response', 'content', 'text', 'json'):
+        raise ValueError
 
     response = session.request(verb, path, **kwargs)
 
@@ -761,19 +799,27 @@ def request(verb, path, session=None, raw=False, **kwargs):
         raise HTTPError(response.url, response.status_code, response.text,
                         response.headers, None)
 
-    try:
-        if raw:
-            return response.json()
-        else:
+    # Return the raw response if requested
+    if format == 'response':
+        return response
+    elif format == 'json':
+        return response.json()
+    elif format == 'text':
+        return response.text
+    elif format == 'content':
+        return response.content
+    else:
+        try:
             obj = _unwrap(response.json())
 
             # ETag is required to update any object
-            # May not be returned on all responses (e.g. listing multiple objects)
+            # May not be returned on all responses (e.g. listing
+            # multiple objects)
             if isinstance(obj, RestObj):
-                setattr(obj, '_headers', response.headers)
+                obj._headers = response.headers
             return obj
-    except ValueError:
-        return response.text
+        except ValueError:
+            return response.text
 
 
 def get_link(obj, rel):